ui: extract components from modalLocalities
Release note: None
